pyproject.toml
[project]
name = "cerebra"
version = "0.1.0"
description = "Terminal-based brain agent CLI with SOUL, MEMORY, USER, TOOLS"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Cerebra contributors" }]
keywords = ["ai", "agent", "brain", "cli", "terminal"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
chromadb = [
    "chromadb>=0.4.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[project.scripts]
cerebra = "cerebra.cli.commands:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cerebra"]

[tool.hatch.build.targets.wheel.sources]
"cerebra" = "cerebra"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
